refactor: split guest artifact snapshot modules
